Moved gameData seeding logic out of startup.cs
