+++ ArchaicQuestII.API/Program.cs
@@ -1,15 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore;
+﻿using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 using Serilog;
-using Serilog.Formatting.Compact;
+using System;
+using System.IO;
 
 namespace ArchaicQuestII.API
 {
@@ -62,7 +57,7 @@ namespace ArchaicQuestII.API
                .UseStartup<Startup>()
                .Build();
             }
-          
+
 
 
             return WebHost.CreateDefaultBuilder(args)
