+++ ArchaicQuestII.GameLogic/Character/Status/Status.cs
@@ -24,78 +24,6 @@ namespace ArchaicQuestII.GameLogic.Character.Status
             Floating = 1 << 9,
             Mounted = 1 << 10,
             Stunned = 1 << 11
-
-        }
-
-
-        public List<Option> SeedData()
-        {
-            var seedData = new List<Option>()
-            {
-                new Option()
-                {
-                    Name = "Sitting",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Standing",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Sleeping",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Fighting",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Resting",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Incapitated",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Dead",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Ghost",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Busy",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Floating",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Mounted",
-                    CreatedBy = "Malleus"
-                },
-                new Option()
-                {
-                    Name = "Stunned",
-                    CreatedBy = "Malleus"
-                }
-
-            };
-
-            return seedData;
         }
     }
 }
